Add clean up worker to delete expired data and refactor
